Fixed Perlin Noise, works well now
